+++ recipes/default.rb
@@ -11,14 +11,15 @@ elsif platform_family?('rhel')
 	include_recipe "yum-epel"
 end
 
-chef_gem 'semantic'
-require 'semantic'
-require 'semantic/core_ext'
-
 include_recipe "openstudio::ruby" unless node[:openstudio][:skip_ruby_install]
 
 # override version of energyplus based on the openstudio version
-if node[:openstudio][:version].to_version >= "1.3.2".to_version
+if Chef::VersionConstraint.new(">= 1.5.4").include?(node[:openstudio][:version])
+  node.default[:energyplus][:version] = "820009"
+  node.default[:energyplus][:long_version] = "8.2.0"
+  node.default[:energyplus][:git_tag] = "v8.2.0-Update-1.2"
+  node.default[:energyplus][:sha] = "8397c2e30b"
+elsif Chef::VersionConstraint.new(">= 1.3.2").include?(node[:openstudio][:version])
   node.default[:energyplus][:version] = "810009"
   node.default[:energyplus][:long_version] = "8.1.0"
 else